css-transforms-1/transform_translate_second_omited.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate with second translation value omited</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="Check if transform supports translate(100px)"> 
    <script type="text/javascript" src="../../resources/testharness.js"></script>
    <script type="text/javascript" src="../../resources/testharnessreport.js"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(100px)";
        var value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText;
        assert_equals(value, "translate(100px)")
    }, "transform_translate_100px");
    </script>    
</body>
</html>
